// app/services/api/models/Post.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Post = {
    _id: string;
    userId: string;
    path: string;
    filename: string;
    mimetype: string;
    size: number;
    position: number;
    tags: Array<string>;
    viewsCount: number;
    url: string;
    sharableUrl: string;
    createdAt: string;
    updatedAt: string;
};